+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.3.3 (SNAPSHOT)
-*
+* Fix config value handling when dsv is x.y.z with undefined x.y but z exists
 
 ### 0.3.2
 * Send target no longer creates base directory